requirements.txt
pyyaml
requests
BeautifulSoup